fixed remaining elements to be able to pull data from Duke
Data was successfully pulled in to HA after this update.

internal_user_id: changed key from cdp_internal_user_id to internalUserID

mail: changed key from email to loginEmailAddress

get_energy_usage: this was the main part needing updating as the data
returned from Duke is different. It appears previously energy and
temperature data was returned in different elements. Now, it's all in
one array. The energy used values are strings in the data and had to be
converted to floats when stored in the returned data.